+++ src/Databricks/Databricks.Autorest/custom/New-AzDatabricksWorkspaceProviderAuthorizationObject.ps1
@@ -21,12 +21,12 @@ Create an in-memory object for WorkspaceProviderAuthorization.
 Create an in-memory object for WorkspaceProviderAuthorization.
 
 .Outputs
-Microsoft.Azure.PowerShell.Cmdlets.Databricks.Models.Api20230201.WorkspaceProviderAuthorization
+Microsoft.Azure.PowerShell.Cmdlets.Databricks.Models.Api20240501.WorkspaceProviderAuthorization
 .Link
 https://learn.microsoft.com/powershell/module/Az.Databricks/new-AzDatabricksWorkspaceProviderAuthorizationObject
 #>
 function New-AzDatabricksWorkspaceProviderAuthorizationObject {
-    [OutputType('Microsoft.Azure.PowerShell.Cmdlets.Databricks.Models.Api20230201.WorkspaceProviderAuthorization')]
+    [OutputType('Microsoft.Azure.PowerShell.Cmdlets.Databricks.Models.Api20240501.WorkspaceProviderAuthorization')]
     [CmdletBinding(PositionalBinding=$false)]
     Param(
 
@@ -39,7 +39,7 @@ function New-AzDatabricksWorkspaceProviderAuthorizationObject {
     )
 
     process {
-        $Object = [Microsoft.Azure.PowerShell.Cmdlets.Databricks.Models.Api20230201.WorkspaceProviderAuthorization]::New()
+        $Object = [Microsoft.Azure.PowerShell.Cmdlets.Databricks.Models.Api20240501.WorkspaceProviderAuthorization]::New()
 
         if ($PSBoundParameters.ContainsKey('PrincipalId')) {
             $Object.PrincipalId = $PrincipalId