+++ src/Resources/Policy.Autorest/custom/Update-AzPolicyDefinition.ps1
@@ -286,6 +286,18 @@ process {
         $calledParameters.Parameter = ConvertTo-Json $existing.Parameter -Depth 100
     }
 
+    if (!$calledParameters.DisplayName) {
+        $calledParameters.DisplayName = $existing.DisplayName
+    }
+
+    if (!$calledParameters.Description) {
+        $calledParameters.Description = $existing.Description
+    }
+
+    if (!$calledParameters.Metadata) {
+        $calledParameters.Metadata = $existing.Metadata
+    }
+
     if ($BackwardCompatible) {
         $calledParameters.BackwardCompatible = $true
     }