+++ src/CosmosDB/CosmosDB/ChangeLog.md
@@ -19,6 +19,10 @@
 -->
 
 ## Upcoming Release
+* Fixed secrets exposure in example documentation.
+
+## Version 1.14.4
+* Fixed the issue that Azure.Core.AccessToken is used before assigned.
 
 ## Version 1.14.3
 * Removed the out-of-date breaking change message for `Get-AzCosmosDBAccountKey`.