+++ src/Storage/Storage.Management/ChangeLog.md
@@ -18,12 +18,43 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Supported account tier Cold
+    - `New-AzStorageAccount`
+    - `Set-AzStorageAccount`
+* Updated list share output display format 
+    - `Get-AzStorageShare`
+* Added warnings for upcoming breaking changes in File cmdlets for removing references to "Microsoft.Azure.Storage.File"
+    - `Close-AzStorageFileHandle`
+    - `Get-AzStorageFile`
+    - `Get-AzStorageFileContent`
+    - `Get-AzStorageFileCopyState`
+    - `Get-AzStorageFileHandle`
+    - `Get-AzStorageShare`
+    - `New-AzStorageDirectory`
+    - `New-AzStorageFileSASToken`
+    - `New-AzStorageShare`
+    - `New-AzStorageShareSASToken`
+    - `Remove-AzStorageDirectory`
+    - `Remove-AzStorageFile`
+    - `Remove-AzStorageShare`
+    - `Set-AzStorageFileContent`
+    - `Set-AzStorageShareQuota`
+    - `Start-AzStorageFileCopy`
+    - `Stop-AzStorageFileCopy`
+
+## Version 7.2.0
+* Upgraded Microsoft.Azure.Storage.DataMovement to 2.0.5
+
+## Version 7.1.0
+* Fixed the issue that Azure.Core.AccessToken is used before assigned.
 * Supported TLS1_3 when creating and updating a storage account 
     - `New-AzStorageAccount`
     - `Set-AzStorageAccount`
 * Fixed sync copy blob issue with -AsJob is specified [#25105]
     - `Copy-AzStorageBlob`
 * Updated Storage.Management.Sdk to support API version 2023-05-01
+* Updated 2 help examples of create storage account cmdlet, with MinimumTlsVersion as TLS1_2.
+    - `New-AzStorageAccount`
 
 ## Version 7.0.0
 * Added a prompt that needs confirmation when upgrading a storage account from StorageV1 or BlobStorage to StorageV2. Can be suppressed with -Force.