+++ src/Synapse/Synapse/ChangeLog.md
@@ -19,6 +19,11 @@
 -->
 
 ## Upcoming Release
+* Fixed secrets exposure in example documentation.
+
+## Version 3.0.9
+* Updated Azure.Analytics.Synapse.Artifacts to 1.0.0-preview.20.
+* Fixed the issue that Azure.Core.AccessToken is used before assigned.
 
 ## Version 3.0.8
 * Upgraded `Microsoft.DataTransfer.Gateway.Encryption` to `5.29.8499.2`