+++ src/Workloads/Workloads/ChangeLog.md
@@ -18,6 +18,9 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Fixed secrets exposure in example documentation.
+
+## Version 0.2.0
 * Added trusted access parameter in Create and Register VIS.
 
 ## Version 0.1.2